Added move history support to game state.
Added a new `moveHistory` property to the connect four game state
that keeps track of each move that is made during the game. This can be
used to replay moves that were made and to rebuild the game state with
the exact moves that were made.

:books: Updated README documentation to have the correct examples representing
the board state and added details about the new `moveHistory` field.

:white_check_mark: Added new unit tests for the new field and updated the
existing tests.

Signed-off-by: Scott Westover <scottwestover2006@gmail.com>